Decryption blobs removed! (test)
Previous commit `bdb469a444e8aa4fca45ffe95a6e4c723230c965` wasnt helped booting the build, it is keep rebooting from the device logo screen. So this time, removing the blobs regarding decryption and try to boot it with CRYPTO flags enabled. So we can assume that the twrp is booting to its spalsh screen or not.